% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{angle}
\alias{angle}
\title{Calculates the angle between two geopositions with respect to an axis.}
\usage{
angle(from_lng, from_lat, to_lng, to_lat, respecto = 0)
}
\arguments{
\item{from_lng}{Origin longitude}

\item{from_lat}{Origin latitude}

\item{to_lng}{Final longitude}

\item{to_lat}{Final latitude}

\item{respecto}{Axis where the angles start to be measured, by default is x-axis with angle 0 (respecto=0)}
}
\value{
angle in degrees
}
\description{
Calculates the angle between two geopositions with respect to an axis.
}
\examples{
x=-96.9236
y=-19.527196
X=-96.91404
Y=-19.5272
angle(x,y,X,Y,respecto=90) #axis y in 90 degrees
}
