% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_color.R
\name{angle_color}
\alias{angle_color}
\title{angle_color}
\usage{
angle_color(angle)
}
\arguments{
\item{angle}{angle from 0 to 360 degrees}
}
\value{
hex color
}
\description{
angle_color
}
\examples{
angle_color(221)
}
