% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citystate.R
\name{citystateServer}
\alias{citystateServer}
\alias{citystateInput}
\alias{citystateOutput}
\alias{citystateApp}
\title{Shiny App of cities with HOLC data}
\usage{
citystateServer(id)

citystateInput(id)

citystateOutput(id)

citystateApp(title = "Redlining")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{title}{title of app}
}
\value{
reactive app

reactive input

reactive input

reactive app
}
\description{
For now, this uses the default redlining GeoJSON URL to extract
cities from user-selected state(s). It is possible to put in your
own URL, but it is tricky to find these URLs.
}
