% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-2.R
\docType{data}
\name{countries_sp}
\alias{countries_sp}
\title{Countries of the world SpatialPolygons object}
\format{SpatialPolygons object with 177 features.}
\source{
\url{https://github.com/ropenscilabs/rnaturalearth}

\url{http://naturalearthdata.com}
}
\usage{
countries_sp
}
\description{
Polygons for countries of the world. The polygons come straight from
\code{rnaturalearth::ne_countries()}.
}
\seealso{
\code{\link{countries_spdf}}, \code{\link{tiny_countries_spdf}}
}
\keyword{datasets}

