% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-2.R
\docType{data}
\name{countries_spdf}
\alias{countries_spdf}
\title{Countries of the world SpatialPolygonsDataFrame object}
\format{SpatialPolygonsDataFrame object with 177 features and 6 variables:
\describe{
 \item{name}{Country name.}
 \item{iso_a3}{Three letter country code, see: \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}.}
 \item{population}{Population of country.}
 \item{gdp}{Gross domestic product of country.}
 \item{region}{Region, one of: Asia, Africa, Europe, Americas, Antarctica, Seven seas (open ocean), Oceania.}
 \item{subregion}{Sub-region.}
}}
\source{
\url{https://github.com/ropenscilabs/rnaturalearth}

\url{http://naturalearthdata.com}
}
\usage{
countries_spdf
}
\description{
Polygons for countries of the world along with data on basic properties for
each country. The polygons and data come from
\code{rnaturalearth::ne_countries()}, with some minimal processing to select
and tidy some variables.
}
\seealso{
\code{\link{countries_sp}}, \code{\link{tiny_countries_spdf}}
}
\keyword{datasets}

