% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mask_and_plot.R
\name{create_mask_and_plot}
\alias{create_mask_and_plot}
\title{Map NDVI per HOLC grade}
\usage{
create_mask_and_plot(
  redlining_sf,
  background_raster = ndvi$raster,
  roads = NULL,
  rivers = NULL
)
}
\arguments{
\item{redlining_sf}{}

\item{background_raster}{}

\item{roads}{}

\item{rivers}{}
}
\value{
list with `plot`, `runtime`, `mask_sf`
}
\description{
Map NDVI per HOLC grade
}
