% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordclouds_by_grade.R
\name{create_wordclouds_by_grade}
\alias{create_wordclouds_by_grade}
\title{Create word cloud per grade}
\usage{
create_wordclouds_by_grade(
  sf_object,
  title = "Healthy food place names word cloud",
  max_size = 25,
  col_select = "name",
  min_word = 3
)
}
\arguments{
\item{sf_object}{object with class `sf`}

\item{title}{title of wordcloud}

\item{max_size}{maximum size for wordcloud plot}

\item{col_select}{column selected from `sf_object`}

\item{min_word}{minimum length for word}
}
\value{
ggplot object
}
\description{
Create word cloud per grade
}
