% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.haversine.R
\name{dt.haversine}
\alias{dt.haversine}
\title{Calculates the separation distance between two geopoints considering the earth curvature.}
\usage{
dt.haversine(lat_from, lon_from, lat_to, lon_to, r = 6378137)
}
\arguments{
\item{lat_from}{Origin lat}

\item{lon_from}{Origin lng}

\item{lat_to}{final lat}

\item{lon_to}{final lng}

\item{r}{Default for haversine distance}
}
\value{
Separation in km
}
\description{
Calculates the separation distance between two geopoints considering the earth curvature.
}
\examples{
dt.haversine(19.2313,-95.12123,18.312312,-96.23123)
}
