% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bbox.R
\name{filter_bbox}
\alias{filter_bbox}
\title{Filter geopoints that fall in a bounding box}
\usage{
filter_bbox(data, x, y, X, Y)
}
\arguments{
\item{data}{data frame with lng and lat colunm names}

\item{x}{paraneter lng1 of bounding box}

\item{y}{paraneter lat1 of bounding box}

\item{X}{paraneter lng2 of bounding box}

\item{Y}{paraneter lat2 of bounding box}
}
\value{
Filter geopoints
}
\description{
Filter geopoints that fall in a bounding box
}
\examples{
example
}
