% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georadious.R
\name{georadius}
\alias{georadius}
\title{georadius}
\usage{
georadius(x, y, radius, nPoints)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}

\item{radius}{radius in km}

\item{nPoints}{number of points over a circle}
}
\value{
data frame
}
\description{
georadius
}
\examples{
geospatial::georadius(x=-96.9236,y=-19.527196,radius = 1,nPoints = 100)

}
