% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_places_student.R
\name{get_places_student}
\alias{get_places_student}
\title{Get Points-of-Interest from city of interest}
\usage{
get_places_student(polygon_layer, type = "food")
}
\arguments{
\item{polygon_layer}{}

\item{type}{type limited to "roads" and "rivers"}
}
\value{
nothing returned; side effect of base plot or message
}
\description{
Get Points-of-Interest from city of interest
}
