% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser.R
\name{geyserServer}
\alias{geyserServer}
\alias{geyserInput}
\alias{geyserUI}
\alias{geyserOutput}
\alias{geyserApp}
\title{Shiny Server for geyser Example}
\usage{
geyserServer(id)

geyserInput(id)

geyserUI(id)

geyserOutput(id)

geyserApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for geyser Example

Shiny Module Input for Geyser

Shiny Module UI for Geyser

Shiny Module Output for Geyser

Shiny Module App for Geyser
}
