% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconsted_route.R
\name{\%do\%}
\alias{\%do\%}
\title{reconstred_route reconstruccion de rutas con geoposicoones en (lon,lat)}
\usage{
obj \%do\% ex
}
\arguments{
\item{data}{}
}
\value{
polyline
}
\description{
reconstred_route reconstruccion de rutas con geoposicoones en (lon,lat)
}
