% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-4.R
\docType{data}
\name{income_grid}
\alias{income_grid}
\title{Gridded income in New York}
\format{SpatialGridDataFrame with 1920 rows and 2400 columns.}
\source{
Seirup, L., and G. Yetman. 2006. U.S. Census Grids (Summary File 3), 2000: Metropolitan Statistical Areas. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). \url{http://dx.doi.org/10.7927/H4Z31WJ0.}

\url{http://sedac.ciesin.columbia.edu/data/set/usgrid-summary-file3-2000-msa}
}
\usage{
income_grid
}
\description{
Gridded median income for New York. Data obtained from NASA Socioeconomic Data and Applications  Center (SEDAC).  Median income comes from the U.S. Census Grids (Summary File 3: Metropolitan Statistical Areas, v1 (2000)) for New York--Northern New Jersey--Long Island, NY--NJ--CT--PA CMSA.
}
\examples{
\dontrun{
# To emulate reading in a raster file as in Exercise 3 in Chapter 4
# This creates the nyc_grid_data/ directory in your current working directory
# and contains the neccessary raster file.
dir.create("nyc_grid_data")
rgdal::writeGDAL(income_grid, "nyc_grid_data/m5602ahhi00.tif", drivername = "GTiff")
}
}
\seealso{
\code{\link{neighborhoods}}, \code{\link{nyc_income}}, \code{\link{water}}
}
\keyword{datasets}

