% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPolypoints.R
\name{InPolypoints}
\alias{InPolypoints}
\title{InPolypoints}
\usage{
InPolypoints(shapefile, nPoints, type)
}
\arguments{
\item{shapefile}{A polygon file}

\item{nPoints}{Number of points inside the polygon}

\item{type}{character; "random" for completely spatial random;
"regular" for regular (systematically aligned) sampling;
"stratified" for stratified random (one single random location in each "cell");
"nonaligned" for nonaligned systematic sampling (nx random y coordinates, ny random x coordinates);
 "hexagonal" for sampling on a hexagonal lattice;
 "clustered" for clustered sampling;
 "Fibonacci" for Fibonacci sampling on the sphere (see references).}
}
\value{
Spatial point
}
\description{
InPolypoints
}
