% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-3.R
\docType{data}
\name{land_cover}
\alias{land_cover}
\title{Global Land Cover}
\format{An object of class SpatialGridDataFrame with 583200 cells and 4 variables.}
\source{
tmap package \url{https://CRAN.R-project.org/package=tmap}
}
\usage{
land_cover
}
\description{
Data directly from the tmap package using \code{data(land, package = "tmap")}.
See \code{\link[tmap]{land}} for details, but note in particular that publication of these maps is only allowed when cited to Tateishi et al. (2014), and when "Geospatial Information Authority of Japan, Chiba University and collaborating organizations." is shown. See \url{http://www.iscgm.org/gm/glcnmo.html#use}."
}
\keyword{datasets}

