% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxigraphComponet.R
\name{maxigraphComponet}
\alias{maxigraphComponet}
\title{Given the disconnected components of a network, the network returns with connections}
\usage{
maxigraphComponet(g)
}
\arguments{
\item{g}{igrap network}
}
\value{
igrap network
}
\description{
Given the disconnected components of a network, the network returns with connections
}
\examples{
make example
}
