% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-3.R
\docType{data}
\name{migration}
\alias{migration}
\title{U.S Net Migration}
\format{RasterLayer with 49 rows and 116 columns on one variable:
\describe{
  \item{net_migration}{estimated number of people who left the grid
  cell between 1990 and 2000.}
  }}
\source{
de Sherbinin, A., M. Levy, S. Adamo, K. MacManus, G. Yetman, V. Mara, L. Razafindrazay, B. Goodrich, T. Srebotnjak, C. Aichele, and L. Pistolesi. 2015. Global Estimated Net Migration Grids by Decade: 1970-2000. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). \url{http://dx.doi.org/10.7927/H4319SVC}. Accessed 7 12 2016.

de Sherbinin, A., M. Levy, S. Adamo, K. MacManus, G. Yetman, V. Mara, L. Razafindrazay, B. Goodrich, T. Srebotnjak, C. Aichele, and L. Pistolesi. 2012. Migration and Risk: Net Migration in Marginal Ecosystems and Hazardous Areas. Environmental Research Letters 7(4): 045602. \url{http://iopscience.iop.org/1748-9326/7/4/045602}.

\url{http://sedac.ciesin.columbia.edu/data/set/popdynamics-global-est-net-migration-grids-1970-2000}
}
\usage{
migration
}
\description{
Gridded estimates of migration between 1990 and 2000 from
NASA Socioeconomic Data and Applications Center (SEDAC).
The global data has been cropped to cover most of the contiguous US.
}
\keyword{datasets}

