% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-4.R
\docType{data}
\name{neighborhoods}
\alias{neighborhoods}
\title{Neighborhoods of NYC}
\format{SpatialPolygonsDataFrame with 195 features.  Each polygon describes an neighborhood tabulation area.}
\source{
Shapefile downloaded from NYC Open Data: \url{https://www1.nyc.gov/site/planning/data-maps/open-data.page}
}
\usage{
neighborhoods
}
\description{
New York City Neighborhood Tabulation Areas.
}
\examples{
\dontrun{
# To emulate reading in a shapefile as in Exercise 2 in Chapter 4
# This creates the nynta_16c/ directory in your current working directory
# and contains the neccessary shapefiles.
rgdal::writeOGR(neighborhoods, "nynta_16c/", "nynta", driver="ESRI Shapefile")
}
}
\seealso{
\code{\link{neighborhoods}}, \code{\link{water}}, \code{\link{income_grid}}
}
\keyword{datasets}

