% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodedge_list2igraph.R
\name{nodedge_list2igraph}
\alias{nodedge_list2igraph}
\title{nodedge_list2igraph}
\usage{
nodedge_list2igraph(
  nodelist,
  edgelist,
  weight = NULL,
  eadf = NULL,
  Directed = FALSE
)
}
\arguments{
\item{Directed}{}
}
\value{

}
\description{
nodedge_list2igraph
}
