% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-4.R
\docType{data}
\name{nyc_income}
\alias{nyc_income}
\title{Median Income in NYC census tracts}
\format{data.frame with 288 rows and 6 variables:
\describe{
 \item{name}{Census tract name.}
 \item{state}{State.}
 \item{county}{County.}
 \item{tract}{Census tract ID.}
 \item{estimate}{Estimate of median income.}
 \item{se}{Standard error on the estimate of median income.}
  }}
\source{
acs pacakge \url{https://CRAN.R-project.org/package=acs}

\url{https://www.census.gov/programs-surveys/acs/}
}
\usage{
nyc_income
}
\description{
Income data comes from the American Community Survey via the \code{acs} package. Specfically, the ACS 5-year estimates ending in 2014, for Table B19013: Median Household Income in the Past 12 Months (in 2014 Inflation-Adjusted Dollars), for tracts in New York County.
}
\seealso{
\code{\link{neighborhoods}}, \code{\link{water}}, \code{\link{income_grid}}
}
\keyword{datasets}

