% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator with which you may pipe a value forward into an expression or function call.}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ecointeraction also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
