% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-3.R
\docType{data}
\name{pop}
\alias{pop}
\title{Population across North East USA}
\format{RasterLayer with 480 rows and 660 columns on one variable: \code{num_people}
the estimated absolute number of people living in the grid cell.}
\source{
Seirup, L., and G. Yetman. 2006. U.S. Census Grids (Summary File 1), 2000. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). \url{http://dx.doi.org/10.7927/H4B85623}. Accessed 7 12 2016.

\url{http://sedac.ciesin.columbia.edu/data/set/usgrid-summary-file1-2000}
}
\usage{
pop
}
\description{
Gridded population data obtained from NASA Socioeconomic Data and Applications
Center (SEDAC).  Population comes from the U.S. Census Grids (Summary File 1), 2000
subsetted to the North East of the U.S.A (latitude within (39, 43) and
and longitude within (-75, -69.5)).
}
\seealso{
\code{\link{pop_by_age}}, \code{\link{prop_by_age}}
}
\keyword{datasets}

