% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-1.R
\docType{data}
\name{preds}
\alias{preds}
\title{Predicted prices for house sales in Corvallis}
\format{Data frame with 1600 rows and 3 variables:
\describe{
 \item{lon, lat}{Longitude and latitude or predicted sale.}
 \item{predicted_price}{Predicted sales price USD.}
}}
\source{
\url{https://www.co.benton.or.us/assessment/page/property-sales-data}
}
\usage{
preds
}
\description{
Predicted sales price over a spatial grid for a sale that satisfies:
\itemize{
  \item{class = "Dwelling",}
  \item{bedrooms = 3,}
  \item{full_bathrooms = 2,}
  \item{finished_squarefeet = 1400,}
  \item{unfinished_squarefeet = 1800}
}
}
\details{
The prediction is based on a regression model with linear terms for
\code{finished_squarefeet}, \code{class}, \code{bedrooms}, \code{full_baths},
 \code{half_baths}, \code{total_squarefeet} and \code{condition}, an
 interaction between \code{finished_squarefeet} and \code{class} and a smooth
  spatial term (using \code{mgcv::s()}).
}
\seealso{
\code{\link{sales}}, \code{\link{ward_sales}}
}
\keyword{datasets}

