% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_plot_sf_layers.R
\name{process_and_plot_sf_layers}
\alias{process_and_plot_sf_layers}
\title{Map an amenity over each grade}
\usage{
process_and_plot_sf_layers(layer1, layer2, output_file = "output_plot.png")
}
\arguments{
\item{layer1}{}

\item{layer2}{}

\item{output_file}{}
}
\value{
list with `ggplot` object and polygon `sf` object
}
\description{
Map an amenity over each grade
}
