% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_city_inventory_data.R
\name{process_city_inventory_data}
\alias{process_city_inventory_data}
\title{Map Denver City provided data per HOLC grade}
\usage{
process_city_inventory_data(
  address,
  inner_file,
  polygon_layer,
  output_filename,
  variable_label = "Tree Density"
)
}
\arguments{
\item{address}{}

\item{inner_file}{}

\item{polygon_layer}{}

\item{output_filename}{}

\item{variable_label}{}
}
\value{
list with `ggplot` object and trees `sf` object
}
\description{
Map Denver City provided data per HOLC grade
}
