% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_city_inventory_data_choice.R
\name{process_city_inventory_data_choice}
\alias{process_city_inventory_data_choice}
\title{Choose dataset and pass patial polygon for analysis}
\usage{
process_city_inventory_data_choice(
  choice,
  polygon_layer,
  source = "https://www.denvergov.org/media/gis/DataCatalog"
)
}
\arguments{
\item{choice}{}

\item{polygon_layer}{}

\item{source}{}
}
\value{
list with `ggplot` object and trees `sf` object
}
\description{
The function process_city_inventory_data_choice allows users to select from a predefined set of datasets for processing. It takes two arguments: choice, an integer that specifies the dataset to process, and polygon_layer, an sf object that represents the geographic area to be analyzed. The choice argument should be a number between 1 and 7, each corresponding to different types of city data:
\enumerate{
  \item **Tree Density** - Tree inventory data.
  \item **Traffic Accidents Density** - Traffic accidents data.
  \item **Instream Sampling Sites Density** - Environmental sampling sites data.
  \item **Soil Samples Density** - Soil sample data.
  \item **Public Art Density** - Public art locations.
  \item **Liquor Licenses Density** - Liquor license data.
  \item **Crime Density** - City crime data.
}
}
