% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress Bar}
\usage{
progress(messagename = "", functionname, ...)
}
\arguments{
\item{messagename}{}

\item{functionname}{}

\item{...}{}
}
\value{
return from call to `functionname` with arguments `...`
}
\description{
Progress Bar
}
