% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-3.R
\docType{data}
\name{prop_by_age}
\alias{prop_by_age}
\title{Population proportions across North East USA by age group}
\format{RasterStack with 480 rows and 660 columns, and 7 layers:
\describe{
  \item{under_1}{estimated proportion of people in grid cell that are less
    than 1 year of age.}
  \item{age_1_4}{estimated proportion of people in grid
    cell that are between 1 and 4 years of age in the grid cell.}
  \item{age_5_17}{estimated proportion of people in grid cell that are
    between 5 and 17 years of age in the grid cell.}
  \item{age_18_24}{estimated
    proportion of people in grid cell that are between 18 and 24 years of age
    in the grid cell.}
  \item{age_25_64}{estimated proportion of people in grid
    cell that are between 24 and 64 years of age in the grid cell.}
  \item{age_65_79}{estimated proportion of people in grid cell that are
    between 65 and 79 years of age in the grid cell.}
  \item{age_80_over}{estimated proportion of people in grid cell that are
    over 80 years of age in the grid cell.}
  }}
\source{
Seirup, L., and G. Yetman. 2006. U.S. Census Grids (Summary File 1), 2000. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). \url{http://dx.doi.org/10.7927/H4B85623}. Accessed 7 12 2016.

\url{http://sedac.ciesin.columbia.edu/data/set/usgrid-summary-file1-2000}
}
\usage{
prop_by_age
}
\description{
Gridded population data obtained from NASA Socioeconomic Data and Applications Center (SEDAC).  Population comes from the U.S. Census Grids (Summary File 1), 2000 subsetted to the North East of the U.S.A (latitude within (39, 43) and longitude within (-75, -69.5)). Absolute counts for each age group were converted to proportions by dividing each age group count by the sum of the counts across age group, on a per grid cell basis.
}
\seealso{
\code{\link{pop}}, \code{\link{pop_by_age}}
}
\keyword{datasets}

