% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recons_inout.R
\name{recons_inout}
\alias{recons_inout}
\title{It joins several gepositions by means of the shortest route in a road network both back and forth.}
\usage{
recons_inout(from_lng, from_lat, data, type = c("in", "out"))
}
\arguments{
\item{from_lng}{Origin lng}

\item{from_lat}{Origin lat}

\item{data}{points away from the origin in a data frame with lng inthe firts columns and lat in the second one.}

\item{type}{reconstructed route from poinr A to B (out) or B to A (in)}
}
\value{
sf file
}
\description{
It joins several gepositions by means of the shortest route in a road network both back and forth.
}
\examples{
x=-96.9236
y=-19.527196
data<-geospatial::georadius(x,y,radius = 1,nPoints = 100)
recons_inout(x,y,data,type="out")
}
