% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconsted_route_fn.R
\name{reconsted_route_fn}
\alias{reconsted_route_fn}
\title{reconstred_route reconstruccion de rutas con geoposicoones en (lon,lat)}
\usage{
reconsted_route_fn(route)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
reconstred_route reconstruccion de rutas con geoposicoones en (lon,lat)
}
