% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redline.R
\name{redlineServer}
\alias{redlineServer}
\alias{redlineInput}
\alias{redlineOutput}
\alias{redlineApp}
\title{Shiny App of cities with HOLC data}
\usage{
redlineServer(id)

redlineInput(id)

redlineOutput(id)

redlineApp(title = "Redlining")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{title}{title of app}
}
\value{
reactive app

reactive input

reactive input

reactive app
}
\description{
For now, this uses the default redlining GeoJSON URL to extract
cities from user-selected state(s). It is possible to put in your
own URL, but it is tricky to find these URLs.
}
