% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-1.R
\docType{data}
\name{sales}
\alias{sales}
\title{House sales in Corvallis 2015}
\format{Data frame with 931 observations and 20 variables:
\describe{
 \item{lon, lat}{Longitude and latitude based on Geocoded address.}
 \item{price}{Sales price USD.}
 \item{finished_squarefeet}{Area of building that is finished (i.e. not garage) square feet.}
 \item{year_built}{Year building was built.}
 \item{date}{Date sale was recorded.}
 \item{address}{Street address.}
 \item{city, state, zip}{City, State and Zip code.}
 \item{acres}{Acres of land.}
 \item{num_dwellings}{Number of dwellings on property.}
 \item{class}{Class of property: Dwelling, Mobile Home, Commercial or RES Feature.}
 \item{condition}{Condition of property: "EX" = excellent, "AV" = average, "G" = good, "F" = fair, "VP" = very poor.}
 \item{total_squarefeet}{Area of building including unfinished space.}
 \item{bedrooms}{Number of bedrooms.}
 \item{full_baths}{Number of full bathrooms.}
 \item{half_baths}{Number of half bathrooms.}
 \item{month}{Month of sale.}
 \item{address_city}{Address used for geo-coding.}
}}
\source{
\url{https://www.co.benton.or.us/assessment/page/property-sales-data}
}
\usage{
sales
}
\description{
Sales of houses in Corvallis, OR, USA in 2015 as provided by Benton County.
}
\details{
This is a subset of all sales that:
\itemize{
  \item{occured within a limited city bounding box: Latitude in (44.52538, 44.60366)
  Longitude in (-123.316845, -123.20698),}
  \item{had a sales price greater than $0,}
  \item{had a reported city of Corvallis, and}
  \item{had a reported address.}
}
}
\seealso{
\code{\link{ward_sales}}, \code{\link{preds}}
}
\keyword{datasets}

