% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny Server for geospatial Package}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output, session}{shiny server reactives}
}
\value{
reactive server
}
\description{
Shiny Server for geospatial Package
}
