% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2nodedge_list.R
\name{shp2nodedge_list}
\alias{shp2nodedge_list}
\title{shp2nodedge_list}
\usage{
shp2nodedge_list(
  ntdata,
  ELComputed = FALSE,
  longlat = FALSE,
  Detailed = FALSE,
  ea.prop = NULL
)
}
\arguments{
\item{ea.prop}{}
}
\value{

}
\description{
shp2nodedge_list
}
