% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_plot.R
\name{split_plot}
\alias{split_plot}
\title{Plot the HOLC grades individually}
\usage{
split_plot(sf_data, roads, rivers)
}
\arguments{
\item{sf_data}{}

\item{roads}{}

\item{rivers}{}
}
\value{
ggplot object
}
\description{
Plot the HOLC grades individually
}
