% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.edgelist.R
\name{Update.edgelist}
\alias{Update.edgelist}
\title{CORREGIR}
\usage{
Update.edgelist(Edgelist, edgeid, fromid, toid, ROL, Detailed)
}
\arguments{
\item{Detailed}{}
}
\value{

}
\description{
CORREGIR
}
