% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-4.R
\docType{data}
\name{water}
\alias{water}
\title{Water areas around NYC}
\format{SpatialPolygonsDataFrame with 3508 features.  Each polygon describes an area of water.}
\source{
Extract downloaded from: \url{https://mapzen.com/data/metro-extracts/metro/new-york_new-york/}
}
\usage{
water
}
\description{
Polygons that describe areas of water in and around NYC.  Data come from a Open Street Map extract provided by Mapzen at \url{https://mapzen.com/documentation/metro-extracts/}.  The New York extract is cropped to the area of New York Country census tracts. Small water areas have been discarded.
}
\seealso{
\code{\link{neighborhoods}}, \code{\link{nyc_income}}, \code{\link{income_grid}}
}
\keyword{datasets}

