% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_average_ndvi.R
\name{yearly_average_ndvi}
\alias{yearly_average_ndvi}
\title{Stream year average NDVI}
\usage{
yearly_average_ndvi(
  polygon_layer,
  output_file = "ndvi.png",
  dx = 0.01,
  dy = 0.01
)
}
\arguments{
\item{polygon_layer}{}

\item{output_file}{}

\item{dx}{}

\item{dy}{}
}
\value{
list with `ggplot`, `processing_time`, `raster` objects
}
\description{
Stream year average NDVI
}
