% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_box.R
\name{bounding_box}
\alias{bounding_box}
\title{Given a central geopoint, draw a bounding box with the geopoint in the center}
\usage{
bounding_box(lat, lon, dist, in.miles = TRUE)
}
\arguments{
\item{lat}{Central latitude}

\item{lon}{Central longitude}

\item{dist}{Distance of separation of the central point (in miles or km)}

\item{in.miles}{Bolean value, by default is miles}
}
\value{

}
\description{
Given a central geopoint, draw a bounding box with the geopoint in the center
}
\examples{
bounding_box(lat=19.01212,log=-96.19216,dist =100,in.miles = F)
}
