% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordclouds_by_grade.R
\name{create_wordclouds_by_grade}
\alias{create_wordclouds_by_grade}
\title{Create word cloud per grade}
\usage{
create_wordclouds_by_grade(
  sf_object,
  output_file = "food_word_cloud_per_grade.png",
  title = "Healthy food place names word cloud",
  max_size = 25,
  col_select = "name"
)
}
\arguments{
\item{sf_object}{}

\item{output_file}{}

\item{title}{}

\item{max_size}{}

\item{col_select}{}
}
\value{
ggplot object
}
\description{
Create word cloud per grade
}
