% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_state_list_from_redlining_data.R
\name{get_city_state_list_from_redlining_data}
\alias{get_city_state_list_from_redlining_data}
\title{List cities where HOLC data are available}
\usage{
get_city_state_list_from_redlining_data(
  url =
    file.path("https://raw.githubusercontent.com/americanpanorama/mapping-inequality-census-crosswalk",
    "main/MIv3Areas_2010TractCrosswalk.geojson")
)
}
\arguments{
\item{url}{URL to the GeoJSON data}
}
\value{
object of class `sf`
}
\description{
Function to get a list of unique cities and states from the redlining data.
See <https://github.com/americanpanorama/mapping-inequality-census-crosswalk>
for panorama files.
}
