% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_city_redlining_data.R
\name{load_city_redlining_data}
\alias{load_city_redlining_data}
\title{Stream HOLC data from a city}
\usage{
load_city_redlining_data(
  city_name,
  url = NULL,
  year = c("2010", "2020"),
  graded = TRUE
)
}
\arguments{
\item{city_name}{}

\item{url}{URL for data (redline default if `NULL`)}

\item{year}{year for data extraction (if `url` is `NULL`)}

\item{graded}{limit to graded cities if `TRUE`}
}
\value{
object of class `sf`
}
\description{
Load and filter redlining data by city
}
