% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_city_redlining_data.R
\name{load_city_redlining_data}
\alias{load_city_redlining_data}
\title{Stream HOLC data from a city}
\usage{
load_city_redlining_data(city_name)
}
\arguments{
\item{city_name}{}
}
\value{
object of class `sf`
}
\description{
Load and filter redlining data by city
}
