% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_city_redlining.R
\name{plot_city_redlining}
\alias{plot_city_redlining}
\title{Plot POI over HOLC grades}
\usage{
plot_city_redlining(redlining_data, filename = "redlining_plot.png")
}
\arguments{
\item{redlining_data}{}

\item{filename}{}
}
\value{
ggplot object
}
\description{
Plot POI over HOLC grades
}
