% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_plot_sf_layers.R
\name{process_and_plot_sf_layers}
\alias{process_and_plot_sf_layers}
\title{Map an amenity over each grade}
\usage{
process_and_plot_sf_layers(object, layer2, amenity)
}
\arguments{
\item{object}{sf object}

\item{layer2}{name of amenity}

\item{amenity}{list of amenities}
}
\value{
list with `ggplot` object and polygon `sf` object
}
\description{
Map an amenity over each grade
}
