% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_satellite_data.R
\name{process_satellite_data}
\alias{process_satellite_data}
\title{Stream NDVI data}
\usage{
process_satellite_data(
  polygon_layer,
  start_date,
  end_date,
  assets,
  fps = 1,
  output_file = "anim.gif"
)
}
\arguments{
\item{polygon_layer}{}

\item{start_date}{}

\item{end_date}{}

\item{assets}{}

\item{fps}{}

\item{output_file}{}
}
\value{
`difftime` object
}
\description{
Function to process satellite data based on an SF polygon's extent
}
